+++ src/extension.ts
@@ -14,13 +14,13 @@ export function activate(context: vscode.ExtensionContext) {
 
 		const panel = vscode.window.createWebviewPanel(
 			'autoLabel',
-			'Auto Label',
+			'BaNaNaLabel',
 			vscode.ViewColumn.One,
 			{
-				enableScripts: true
+				enableScripts: true,
 			}
 		);
-
+		panel.iconPath = vscode.Uri.file(path.join(context.extensionPath, 'media', 'icon.png'));
 		if (lastDirectory && fs.existsSync(lastDirectory)) {
 			const files = fs.readdirSync(lastDirectory).filter(file => file.endsWith('.json') || file.endsWith('.xml'));
 			panel.webview.html = getWebviewContent(lastDirectory, files, recentDirectories);
@@ -49,6 +49,11 @@ export function activate(context: vscode.ExtensionContext) {
 			} else if (message.command === 'addLabel') {
 				const { label, files, directory } = message;
 
+				if (!label || files.some((file: { file: string, value: string }) => !file.value)) {
+					vscode.window.showErrorMessage('The label name and value cannot be null.');
+					return;
+				}
+
 				for (const { file, value } of files) {
 					const absoluteFilePath = path.join(directory, file);
 					if (!fs.existsSync(absoluteFilePath)) {
@@ -64,6 +69,10 @@ export function activate(context: vscode.ExtensionContext) {
 							if (fileContent) {
 								jsonContent = JSON.parse(fileContent);
 							}
+							if (jsonContent.hasOwnProperty(label)) {
+								vscode.window.showErrorMessage(`Label "${label}" already exists in file "${absoluteFilePath}" with value "${jsonContent[label]}".`);
+								continue;
+							}
 							jsonContent[label] = value;
 							fs.writeFileSync(absoluteFilePath, JSON.stringify(jsonContent, null, 2), 'utf8');
 						} else if (file.endsWith('.xml')) {
@@ -78,21 +87,26 @@ export function activate(context: vscode.ExtensionContext) {
 							if (!xmlContent.root.data) {
 								xmlContent.root.data = [];
 							}
+							const existingLabel = xmlContent.root.data.find((item: any) => item.$.name === label);
+							if (existingLabel) {
+								vscode.window.showErrorMessage(`Label "${label}" already exists in file "${absoluteFilePath}" with value "${existingLabel.value}".`);
+								continue;
+							}
 							xmlContent.root.data.push({
 								$: { name: label },
 								value: value
 							});
-							const builder = new Builder({ headless: true, renderOpts: { pretty: true, indent: '  ', newline: '\n' } });
+							const builder = new Builder({ headless: false, renderOpts: { pretty: true, indent: '  ', newline: '\n' } });
 							const newXmlContent = builder.buildObject(xmlContent);
+
 							fs.writeFileSync(absoluteFilePath, newXmlContent, 'utf8');
 						}
 					} catch (error) {
 						if (error instanceof Error) {
 							vscode.window.showErrorMessage(`Error processing file "${absoluteFilePath}": ${error.message}`);
+						}
 					}
 				}
-			}
-
 				vscode.window.showInformationMessage(`Label "${label}" added to selected files.`);
 			}
 		});
@@ -112,13 +126,85 @@ function getInitialWebviewContent(recentDirectories: string[]) {
 		<head>
 			<meta charset="UTF-8">
 			<meta name="viewport" content="width=device-width, initial-scale=1.0">
-			<title>Auto Label</title>
+			<title>BaNaNaLabel</title>
+			<script src="https://kit.fontawesome.com/bdac89dd37.js" crossorigin="anonymous"></script>
+			<style>
+			*{
+			font-size: 16px;
+			margin-top: 5px;
+			}
+			body{
+				background-color:black;
+			}
+			#directoryDiv {
+				display: flex;
+				align-items: center;
+			}
+			.dirElem {
+				margin-right: 10px;
+				padding: 5px;
+			}
+		    h1{
+			font-size: 36px;
+			}
+			#folderIcon {
+				margin-left: 5px;
+			}
+           #LblNameDiv{
+			display:flex;
+			flex-direction:column;
+			max-width:200px;
+			}
+			#labelName{
+			max-width: 150px;
+			resize: horizontal;
+			}
+			button,input,select{
+			color: hsla(207, 11%, 39%, 1);;
+			background-color:white;
+			border-color: hsla(207, 11%, 39%, 1);
+			border-style:solid;
+			border-width: 1px;
+			border-radius: 5px;
+			padding: max(5px);
+
+			}
+			select{
+			padding:5px;
+			max-width:140px;
+			}
+			select,input{
+			max-width:80px;
+			}
+			select:hover ,input:hover,button:hover,button:active{
+				background-color:lightgrey;
+				color : black;
+			}
+			#AddBtn{
+			width: 100px;
+		    color:hsla(50, 100%, 52%, 1);
+			border-color:hsla(50, 100%, 52%, 1);
+			background-color:black;
+			}
+			#AddBtn:hover,#AddBtn:active{
+			background-color:hsla(50, 100%, 52%, 1);
+			color:black;
+			border-color: black;
+			}
+			#title{
+			color: hsla(50, 100%, 52%, 1);
+			}
+			input{
+			min-width: 180px
+			max-width: 400px;
+			}
+		</style>
 		</head>
 		<body>
-			<h1>Auto Label</h1>
-			<button onclick="selectDirectory()">Seleziona Directory</button>
+			<h1 id="title">BaNaNaLabel</h1>
+			<button onclick="selectDirectory()" tabindex="-1">Select Directory <i class="fa-solid fa-folder"></i></button>
 			<select id="recentDirectories" onchange="selectRecentDirectory()">
-				<option value="">Seleziona una directory recente</option>
+				<option value="">Recent folders</option>
 				${recentDirsOptions}
 			</select>
 			<script>
@@ -140,6 +226,13 @@ function getInitialWebviewContent(recentDirectories: string[]) {
 						});
 					}
 				}
+
+				document.addEventListener('keydown', function(event) {
+					if (event.key === 'Enter') {
+						event.preventDefault();
+						document.getElementById('AddBtn').click();
+					}
+				});
 			</script>
 		</body>
 		</html>
@@ -148,9 +241,9 @@ function getInitialWebviewContent(recentDirectories: string[]) {
 
 function getWebviewContent(directory: string, files: string[], recentDirectories: string[]) {
 	const fileCheckboxes = files.map(file => `
-		<input type="checkbox" id="${file}" name="files" value="${file}" onchange="toggleInput('${file}')">
+		<input type="checkbox" id="${file}" name="files" value="${file}" onchange="toggleInput('${file}')" onkeydown="handleCheckboxKeydown(event, '${file}')">
 		<label for="${file}">${file}</label>
-		<input type="text" id="value_${file}" name="value_${file}" placeholder="Valore Label" style="display:none;"><br>
+		<input type="text" id="value_${file}" name="value_${file}" placeholder="Label Value" style="display:none; resize: horizontal;" onblur="checkInput('${file}')" onkeydown="handleInputKeydown(event, '${file}')"><br>
 	`).join('');
 
 	const recentDirsOptions = recentDirectories.map(dir => `<option value="${dir}">${dir}</option>`).join('');
@@ -161,22 +254,97 @@ function getWebviewContent(directory: string, files: string[], recentDirectories
 		<head>
 			<meta charset="UTF-8">
 			<meta name="viewport" content="width=device-width, initial-scale=1.0">
-			<title>Auto Label</title>
+			<title>BaNaNaLabel</title>
+			<script src="https://kit.fontawesome.com/bdac89dd37.js" crossorigin="anonymous"></script>
+			<style>
+			*{
+			font-size: 16px;
+			margin-top: 5px;
+			}
+			body{
+				background-color:black;
+			}
+			#directoryDiv {
+				display: flex;
+				align-items: center;
+			}
+			.dirElem {
+				margin-right: 10px;
+				padding: 5px;
+			}
+		    h1{
+			font-size: 36px;
+			}
+			#folderIcon {
+				margin-left: 5px;
+			}
+           #LblNameDiv{
+			display:flex;
+			flex-direction:column;
+			max-width:200px;
+			}
+			#labelName{
+			max-width: 150px;
+			resize: horizontal;
+			}
+			button,input,select{
+			color: hsla(207, 11%, 39%, 1);;
+			background-color:white;
+			border-color: hsla(207, 11%, 39%, 1);
+			border-style:solid;
+			border-width: 1px;
+			border-radius: 5px;
+			padding: max(5px);
+
+			}
+			select{
+			padding:5px;
+			max-width:140px;
+			}
+			select,input{
+			max-width:80px;
+			}
+			select:hover ,input:hover,button:hover,button:active{
+				background-color:lightgrey;
+				color : black;
+			}
+			#AddBtn{
+			width: 100px;
+		    color:hsla(50, 100%, 52%, 1);
+			border-color:hsla(50, 100%, 52%, 1);
+			background-color:black;
+			}
+			#AddBtn:hover,#AddBtn:active{
+			background-color:hsla(50, 100%, 52%, 1);
+			color:black;
+			border-color: black;
+			}
+			#title{
+			color: hsla(50, 100%, 52%, 1);
+			}
+			input{
+			min-width: 180px
+			max-width: 400px;
+			}
+		</style>
 		</head>
 		<body>
-			<h1>Auto Label</h1>
-			<p>Directory selezionata: ${directory}</p>
-			<button onclick="selectDirectory()">Cambia Directory</button>
-			<select id="recentDirectories" onchange="selectRecentDirectory()">
-				<option value="">Seleziona una directory recente</option>
-				${recentDirsOptions}
-			</select>
+			<h1 id="title">BaNaNaLabel</h1>
+			<div id="directoryDiv">
+			<button class="dirElem" onclick="selectDirectory()" tabindex="-1"> ${directory} <i id="folderIcon" class="fa-regular fa-folder"></i></button>
+		<select class="dirElem" id="recentDirectories" onchange="selectRecentDirectory()">
+			<option value="">Recent folders</i></option>
+			${recentDirsOptions}
+		</select>
+		</div>
 			<form id="labelForm">
-				<label for="labelName">Nome Label:</label>
-				<input type="text" id="labelName" name="labelName"><br><br>
-				<label for="files">File influenzati:</label><br>
+			<div id="LblNameDiv">
+				<label  for="labelName">Label name</label>
+				<input type="text" id="labelName" placeholder="LabelName" name="labelName" style="resize: horizontal;"><br><br>
+				</div>
+				<label for="files">File list</label><br>
 				${fileCheckboxes}
-				<button type="button" onclick="addLabel()">Aggiungi Label</button>
+				<button id="AddBtn" type="button" onclick="addLabel()" tabindex="-1">Add Label</button>
 			</form>
 			<script>
 				const vscode = acquireVsCodeApi();
@@ -201,7 +369,52 @@ function getWebviewContent(directory: string, files: string[], recentDirectories
 				function toggleInput(file) {
 					const input = document.getElementById('value_' + file);
 					input.style.display = input.style.display === 'none' ? 'inline' : 'none';
+					if (input.style.display === 'inline') {
+						input.focus();
+					}
 				}
+
+				function checkInput(file) {
+					const input = document.getElementById('value_' + file);
+					const checkbox = document.getElementById(file);
+					if (!input.value) {
+						checkbox.checked = false;
+						input.style.display = 'none';
+					}
+				}
+
+				function handleCheckboxKeydown(event, file) {
+					if (event.key === 'Tab') {
+						event.preventDefault();
+						const input = document.getElementById('value_' + file);
+						const checkbox = document.getElementById(file);
+						if (checkbox.checked) {
+							input.focus();
+						} else {
+							checkbox.checked = true;
+							input.style.display = 'inline';
+							input.focus();
+						}
+					}
+				}
+
+				function handleInputKeydown(event, file) {
+					if (event.key === 'Tab') {
+						const input = document.getElementById('value_' + file);
+						const checkbox = document.getElementById(file);
+						if (!input.value) {
+							checkbox.checked = false;
+							input.style.display = 'none';
+						}
+					}
+				}
+
+				document.addEventListener('keydown', function(event) {
+					if (event.key === 'Enter') {
+						event.preventDefault();
+						document.getElementById('AddBtn').click();
+					}
+				});
 
 				function addLabel() {
 					const label = document.getElementById('labelName').value;
@@ -209,6 +422,14 @@ function getWebviewContent(directory: string, files: string[], recentDirectories
 						file: checkbox.value,
 						value: document.getElementById('value_' + checkbox.value).value
 					}));
+
+					if (!label || files.some(file => !file.value)) {
+						vscode.postMessage({
+							command: 'showError',
+							message: 'The label name and value cannot be null.'
+						});
+						return;
+					}
 
 					vscode.postMessage({
 						command: 'addLabel',
